- Fixed references to software commits in metadata - Fixed ORCID references - Fixed continue simulation
