+++ wrappers/do_sims_onedata.py
@@ -12,46 +12,22 @@
 # additional modules needed
 # apt-get install python3-xattr
 # or yum install -y python36-pyxattr
-# import subprocess, os, shutil, xattr, json, datetime, fcntl, sys, shlex
-import subprocess
 import os
 import xattr
 import json
 import datetime
-import sys
-import shlex
+import shutil
 
 from threading import Thread
 from queue import Queue
 
 #
-from arguments import get_sys_args
+from arguments import get_sys_args, _run_Popen, _run_Popen_interactive
 
 onedataSimPath = os.path.dirname(os.path.abspath(__file__))
 
 
 # ----- utils -----
-def _run_Popen_interactive(command):
-
-    print(command+'\n')
-    p = subprocess.Popen(shlex.split(command), env=os.environ,
-                         stdin=sys.stdin, stdout=sys.stdout,
-                         stderr=sys.stderr)
-    p.wait()
-
-
-def _run_Popen(command, timeout=None):
-    print(command+'\n')
-    p = subprocess.Popen(command + ' 2>&1', shell=True, env=os.environ,
-                         stdin=subprocess.PIPE, stdout=subprocess.PIPE)
-    p.wait(timeout)
-    res = ""
-    if p.returncode != 0:
-        print("Return code: "+str(res)+'\n')
-    else:
-        res = p.communicate()[0]
-    return res
-
 
 def _xsd_dateTime():
 
@@ -86,11 +62,12 @@ def _add_json(j, j_new):
     return _add_json([j], j_new)
 
 def _replace_common_patterns(s, catcodename, arti_params_dict):
-        s = s.replace('CATCODENAME', catcodename)
-        s = s.replace('ORCID', arti_params_dict['u'])
-        # from arguments.py
-        s = s.replace('COMMITSHAARTI', arti_params_dict['priv_articommit'])
-        s = s.replace('COMMITSHAODSIM', arti_params_dict['priv_odsimcommit'])
+    s = s.replace('CATCODENAME', catcodename)
+    s = s.replace('ORCID', arti_params_dict['u'])
+    # private generated by arguments.py
+    s = s.replace('COMMITSHAARTI', arti_params_dict['priv_articommit'])
+    s = s.replace('COMMITSHAODSIM', arti_params_dict['priv_odsimcommit'])
+    return s
 
 #####
 def get_first_catalog_metadata_json(catcodename, arti_params_dict):
@@ -172,7 +149,8 @@ def _get_lst_output_metadata(filecode):
             return s
 
 
-def get_dataset_metadata(catcodename, filecode, startdate, end_date):
+def get_dataset_metadata(catcodename, filecode, startdate, end_date,
+                         arti_params_dict):
 
     mdlistaux = [_get_bin_output_metadata(filecode),
                  _get_lst_output_metadata(filecode),
@@ -188,12 +166,14 @@ def get_dataset_metadata(catcodename, filecode, startdate, end_date):
 
 
 #########
-def _run_check_and_copy_results(catcodename, filecode, task, onedata_path):
+def _run_check_and_copy_results(catcodename, filecode, task, onedata_path,
+                                arti_params_dict):
 
     # check if the results are already in onedata before running the task
     runtask = False
     mdlist_prev = get_dataset_metadata(catcodename, filecode,
-                                       _xsd_dateTime(), _xsd_dateTime())
+                                       _xsd_dateTime(), _xsd_dateTime(),
+                                       arti_params_dict)
     for md in mdlist_prev:
         id = json.loads(md)['@id']
         # We should also check if the existent metadata is well formed
@@ -212,7 +192,8 @@ def _run_check_and_copy_results(catcodename, filecode, task, onedata_path):
             start_date = _xsd_dateTime()
             _run_Popen(task)
             metadatalist = get_dataset_metadata(catcodename, filecode, 
-                                                start_date, _xsd_dateTime())
+                                                start_date, _xsd_dateTime(),
+                                                arti_params_dict)
             
             for md in metadatalist:
                 id = json.loads(md)['@id']
@@ -232,6 +213,10 @@ q = Queue()
 
 
 def _producer(catcodename, arti_params):
+
+    # clean a possible previous simulation
+    if os.path.exists(catcodename):
+        shutil.rmtree(catcodename, ignore_errors=True)
 
     cmd = 'do_sims.sh ' + arti_params
     _run_Popen_interactive(cmd)
@@ -268,12 +253,12 @@ def _producer(catcodename, arti_params):
             q.put((filecode, task))
 
 
-def _consumer(catcodename, onedata_path):
+def _consumer(catcodename, onedata_path, arti_params_dict):
     while True:
         (filecode, task) = q.get()
         try:
             _run_check_and_copy_results(catcodename, filecode, task,
-                                        onedata_path)
+                                        onedata_path, arti_params_dict)
             print('Completed NRUN: ' + str(filecode) + '  ' + task)
             q.task_done()
         except Exception as inst:
@@ -313,7 +298,8 @@ except Exception as inst:
     raise inst
 
 for i in range(int(arti_params_dict["j"])):  # processors
-    t = Thread(target=_consumer, args=(catcodename, onedata_path))
+    t = Thread(target=_consumer, args=(catcodename, onedata_path,
+                                       arti_params_dict))
     t.daemon = True
     t.start()
